# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LevelUp Gamified Self-Improvement Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface RegisterBody {
  /**
   * @minLength 3
   * @maxLength 30
   */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface UserStats {
  /** @minimum 0 */
  strength: number;
  /** @minimum 0 */
  intelligence: number;
  /** @minimum 0 */
  discipline: number;
  /** @minimum 0 */
  health: number;
}

export interface UserProfile {
  id: number;
  username: string;
  email: string;
  level: number;
  xp: number;
  xpToNextLevel: number;
  streak: number;
  stats: UserStats;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export type QuestDifficulty =
  (typeof QuestDifficulty)[keyof typeof QuestDifficulty];

export const QuestDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export type QuestCategory = (typeof QuestCategory)[keyof typeof QuestCategory];

export const QuestCategory = {
  strength: "strength",
  intelligence: "intelligence",
  discipline: "discipline",
  health: "health",
} as const;

export interface Quest {
  id: number;
  userId: number;
  title: string;
  description?: string | null;
  difficulty: QuestDifficulty;
  category: QuestCategory;
  xpReward: number;
  completed: boolean;
  completedAt?: string | null;
  createdAt: string;
}

export type CreateQuestBodyDifficulty =
  (typeof CreateQuestBodyDifficulty)[keyof typeof CreateQuestBodyDifficulty];

export const CreateQuestBodyDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export type CreateQuestBodyCategory =
  (typeof CreateQuestBodyCategory)[keyof typeof CreateQuestBodyCategory];

export const CreateQuestBodyCategory = {
  strength: "strength",
  intelligence: "intelligence",
  discipline: "discipline",
  health: "health",
} as const;

export interface CreateQuestBody {
  /**
   * @minLength 1
   * @maxLength 200
   */
  title: string;
  /** @maxLength 500 */
  description?: string;
  difficulty: CreateQuestBodyDifficulty;
  category: CreateQuestBodyCategory;
}

export type UpdateQuestBodyDifficulty =
  (typeof UpdateQuestBodyDifficulty)[keyof typeof UpdateQuestBodyDifficulty];

export const UpdateQuestBodyDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export type UpdateQuestBodyCategory =
  (typeof UpdateQuestBodyCategory)[keyof typeof UpdateQuestBodyCategory];

export const UpdateQuestBodyCategory = {
  strength: "strength",
  intelligence: "intelligence",
  discipline: "discipline",
  health: "health",
} as const;

export interface UpdateQuestBody {
  completed?: boolean;
  title?: string;
  description?: string;
  difficulty?: UpdateQuestBodyDifficulty;
  category?: UpdateQuestBodyCategory;
}

export interface QuestCompleteResponse {
  quest: Quest;
  user: UserProfile;
  xpGained: number;
  leveledUp: boolean;
  newLevel?: number | null;
  streakUpdated: boolean;
}

export interface StatsSummary {
  totalQuestsCompleted: number;
  totalQuestsPending: number;
  currentStreak: number;
  longestStreak: number;
  xpToNextLevel: number;
  /** 0-100 percentage to next level */
  xpProgress: number;
  totalXpEarned: number;
  questsCompletedToday: number;
}

export interface DailyStats {
  date: string;
  xpEarned: number;
  questsCompleted: number;
  dayLabel: string;
}

export interface LeaderboardEntry {
  rank: number;
  username: string;
  level: number;
  xp: number;
  streak: number;
}

export type GetQuestsParams = {
  completed?: boolean;
};